pyproject.toml
[project]
name = "iot-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "influxdb-client>=1.50.0",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.2",
]
